+++ changes.md
@@ -19,7 +19,7 @@
 13. Introduce direct streamers - like TBasket or TRef
    Benefit - one can add custom streamers of such kind or reuse existing
 14. Handle TMatrixTSym classes in I/O  
-15. Implement for TH3 box2, box3, glbox2 draw options 
+15. Implement for TH3 box2, box3, glbox2, glcol draw options 
 
 
 ## Changes in 5.0.3